Refactor: Apply new module
